Refactor: prompt에서 에러 메시지 출력에 쓰인   Console.print()를 OutputView의 print메서드로 추상화
