+++ src/Controllers/index.js
@@ -1,4 +1,4 @@
-export { default as AnnounceResult } from "./AnnounceResultController";
-export { default as DrawnLottoController } from "./DrawnLottoController";
-export { default as LottoGame } from "./LottoGame";
-export { default as LottoPurchaseController } from "./LottoPurchaseController";
+export { default as AnnounceResult } from "./AnnounceResultController.js";
+export { default as DrawnLottoController } from "./DrawnLottoController.js";
+export { default as LottoGame } from "./LottoGame.js";
+export { default as LottoPurchaseController } from "./LottoPurchaseController.js";
