+++ src/Views/index.js
@@ -1,0 +1,2 @@
+export { default as OutputView } from "./OutputView.js";
+export { default as InputView } from "./InputView.js";
